requirements.txt
flask==2.3.3
werkzeug==2.3.7
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
pillow==10.0.0
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.65.0
opencv-python==4.8.1.78
scipy==1.11.3